FILE: moves file to internal/ directory
