+++ libs/globetrotter/data-access/src/lib/internal/selection.interface.ts
