+++ libs/globetrotter/data-access/src/lib/select.service.ts
@@ -4,7 +4,7 @@ import { map } from 'rxjs/operators';
 import { State } from '@atocha/core/data-access';
 import { QuizType, mapRegionsToPlacesModel } from '@atocha/globetrotter/util';
 import { PlaceService } from './place.service';
-import { Selection, SelectionParams } from './selection.interface';
+import { Selection, SelectionParams } from './internal/selection.interface';
 
 @Injectable({
   providedIn: 'root',
